Moved the build tool specific details of how to run a scenario into separate classes and out of `Main`.
This commit doesn't address the duplication in the build tool specific code, it simply moves the code.
